Sample Code modifications
Hosted iFrame: Contact details mandatory, HPP is now on the right side of the PAX image

Lightbox (modal): modified the form and now the cardholder can enter the address (first you need to enter the card data).

Redirect: changed the background and now the Handpoint logo shows better but the code background is "white".